Make OrderingSearch a different class (rather than subclassing to implement it)
